separate functionality into different files
